[#64]/fix change location of recoil initial state from app to index
