+++ web/src/index.tsx
@@ -3,12 +3,13 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
+import { getInitialUserInfo } from 'recoil/atoms/initializer';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
   <React.StrictMode>
-    <RecoilRoot>
+    <RecoilRoot initializeState={getInitialUserInfo}>
       <BrowserRouter>
         <App />
       </BrowserRouter>
